excluding build for windows syscalls on non-Windows OSes
