+++ cmd.go
@@ -3,8 +3,6 @@ package input_autocomplete
 import (
 	"io/ioutil"
 	"os"
-
-	"golang.org/x/sys/windows"
 )
 
 type Cmd struct{}
@@ -44,15 +42,4 @@ func (c Cmd) IsDir(path string) (bool, error) {
 	}
 
 	return info.IsDir(), nil
-}
-
-func EnableVirtalTerminalWindows() error {
-	var originalMode uint32
-	stdout := windows.Handle(os.Stdout.Fd())
-
-	if err := windows.GetConsoleMode(stdout, &originalMode); err != nil {
-		return err
-	}
-
-	return windows.SetConsoleMode(stdout, originalMode|windows.ENABLE_VIRTUAL_TERMINAL_PROCESSING)
 }
