+++ enable_virtual_terminal_exclude.go
@@ -1,0 +1,10 @@
+// +build !windows
+
+package input_autocomplete
+
+// We make the assumption that on non-Windows OSes support ANSI escape
+// sequences by default.
+
+func EnableVirtalTerminalWindows() error {
+	return nil
+}
